moved check to extra file
